HACK: include/types: Support 64-bit in phys_vec
The vfio's dmabuf uAPI Support 64-bit input but capped to 32-bit, which
failed any case that needs a size larger than 4GB.

+++ include/linux/types.h
@@ -172,7 +172,7 @@ typedef u32 phys_addr_t;
 
 struct phys_vec {
 	phys_addr_t	paddr;
-	u32		len;
+	size_t		len;
 };
 
 typedef phys_addr_t resource_size_t;
